+++ src/shopgate/opencart/database.php
@@ -34,8 +34,7 @@ class ShopgateOpencartDatabase extends ShopgateObject
 
     public function __construct()
     {
-        $database = @($GLOBALS["___mysqli_ston"] = mysqli_connect(DB_HOSTNAME, DB_USERNAME, DB_PASSWORD));
-        ((bool)mysqli_query($database, "USE " . constant('DB_DATABASE')));
+        $database = @($GLOBALS["___mysqli_ston"] = mysqli_connect(DB_HOSTNAME, DB_USERNAME, DB_PASSWORD, DB_DATABASE));
         mysqli_set_charset($database, 'utf8');
 
         $this->_database = $database;